Add <article> element along with content
Add an <article> element to the page and set it as the parent element for all other content of the page.
Within article element, create a header element to represent introductory content of the page.
The header element consists of h1, p, h2 and dl elements to represent the title and other information.